requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
nltk>=3.8.1
spacy>=3.7.2
Pillow>=10.0.0
imgurpython>=1.1.7
notion-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pytesseract>=0.3.10
googletrans>=3.1.0a0 